+++ src/MEGASync/transfers/gui/TransfersSummaryWidget.h
@@ -46,8 +46,8 @@ public:
     void adjustFontSizeToText(QFont *font, int maxWidth, QString uploadText, int fontsize = 12.0);
     int adjustSizeToText(QFont *font, int maxWidth, int minWidth, int margins, long long partial, long long total, int &posDotsPartial, int &posDotsTotal, QString &text, int fontsize);
 
-    void setUploads(long long  completed, long long total);
-    void setDownloads(long long  completed, long long total);
+    void setUploads(uint completed, uint total);
+    void setDownloads(uint completed, uint total);
 
     void initialize();
     void reset();
@@ -159,10 +159,10 @@ private:
 
     qreal speed = 0;
 
-    long long totalUploads;
-    long long currentUpload;
-    long long totalDownloads;
-    long long currentDownload;
+    uint totalUploads;
+    uint currentUpload;
+    uint totalDownloads;
+    uint currentDownload;
 
     QFont fontUploads;
     QFont fontDownloads;
